Implement missing functionality in FairMixedSource
